Refactor IdentifierTypes.
Removed IdentiferTypes from health, stuck it in bika.lims.

Renamed patient/PatientIdentifiers to 'Identifier', this is applied
by code in the bika.lims/adapters/identifiers.py extender.

The keyword index: "Identifiers" will be created.  This index must exist
in all catalogs where objects are registered to contain Identifers of
this type (IHaveIdentifiers).

Or not.  up to you.